refactor: 上提 MtouchHelper 共性逻辑，消除 Minitouch/Maatouch/AndrowsExtras 重复代码
