+++ source/MaaAdbControlUnit/EmulatorExtras/AndrowsExtras.h
@@ -3,12 +3,9 @@
 #ifndef __ANDROID__
 
 #include "Base/UnitBase.h"
-#include "Input/AdbShellInput.h"
 #include "Input/MtouchHelper.h"
-#include "Invoke/InvokeApp.h"
 #include "MaaUtils/NoWarningCVMat.hpp"
 
-#include <filesystem>
 #include <optional>
 #include <string>
 
@@ -39,15 +36,6 @@ public: // from ScreencapBase
     virtual bool init() override;
     virtual std::optional<cv::Mat> screencap() override;
 
-public: // from InputBase (via MtouchHelper)
-    virtual MaaControllerFeature get_features() const override;
-
-    virtual bool click_key(int key) override;
-    virtual bool input_text(const std::string& text) override;
-
-    virtual bool key_down(int key) override;
-    virtual bool key_up(int key) override;
-
 public: // from ControlUnitSink
     virtual void on_app_started(const std::string& intent) override;
     virtual void on_app_stopped(const std::string& intent) override;
@@ -55,30 +43,8 @@ public: // from ControlUnitSink
 
 protected: // from MtouchHelper
     virtual bool request_display_info() override;
-    virtual std::pair<int, int> screen_to_touch(int x, int y) override { return _screen_to_touch(x, y); }
-    virtual std::pair<int, int> screen_to_touch(double x, double y) override { return _screen_to_touch(x, y); }
 
 private:
-    template <typename T1, typename T2>
-    inline std::pair<int, int> _screen_to_touch(T1 x, T2 y)
-    {
-        auto make_pair = [](double x, double y) {
-            return std::make_pair(static_cast<int>(round(x)), static_cast<int>(round(y)));
-        };
-        switch (orientation_) {
-        case 0:
-            return make_pair(x * xscale_, y * yscale_);
-        case 1:
-            return make_pair(touch_height_ - y * yscale_, x * xscale_);
-        case 2:
-            return make_pair(touch_width_ - x * xscale_, touch_height_ - y * yscale_);
-        case 3:
-            return make_pair(y * yscale_, touch_width_ - x * xscale_);
-        default:
-            return make_pair(x * xscale_, y * yscale_);
-        }
-    }
-
     // Query logical display_id via ADB shell, cache result.
     // Returns true if display ID was successfully resolved.
     bool query_display_id();
@@ -92,13 +58,9 @@ private:
     // Reinvoke the already-pushed minitouch without re-pushing the binary.
     // Used when app/display/resolution changes but the binary has not changed.
     bool reinvoke_minitouch();
-    void remove_binary();
 
     // Execute an ADB shell command and return the trimmed output.
     std::optional<std::string> adb_shell(const std::string& cmd);
-
-    std::filesystem::path agent_path_;
-    std::vector<std::string> arch_list_;
 
     std::string app_package_;
 
@@ -111,9 +73,6 @@ private:
 
     // ADB shell command generator (used to query display/event IDs)
     ProcessArgvGenerator adb_shell_argv_;
-
-    std::shared_ptr<InvokeApp> invoke_app_ = std::make_shared<InvokeApp>();
-    std::shared_ptr<AdbShellInput> adb_shell_input_ = std::make_shared<AdbShellInput>();
 };
 
 MAA_CTRL_UNIT_NS_END
