+++ source/MaaAdbControlUnit/Input/MaatouchInput.h
@@ -1,9 +1,5 @@
 #pragma once
 
-#include <filesystem>
-
-#include "Base/UnitBase.h"
-#include "Invoke/InvokeApp.h"
 #include "MtouchHelper.h"
 
 #include "Common/Conf.h"
@@ -14,8 +10,9 @@ class MaatouchInput : public MtouchHelper
 {
 public:
     explicit MaatouchInput(std::filesystem::path agent_path)
-        : agent_path_(std::move(agent_path))
     {
+        agent_path_ = std::move(agent_path);
+        invoke_app_ = std::make_shared<InvokeApp>();
         children_.emplace_back(invoke_app_);
     }
 
@@ -39,23 +36,20 @@ public: // from ControlUnitSink
     virtual void on_image_resolution_changed(const std::pair<int, int>& pre, const std::pair<int, int>& cur) override;
 
 protected: // from MtouchHelper
-    virtual std::pair<int, int> screen_to_touch(int x, int y) override { return _screen_to_touch(x, y); }
+    virtual std::pair<int, int> screen_to_touch(int x, int y) override { return screen_to_touch_no_orient(x, y); }
 
-    virtual std::pair<int, int> screen_to_touch(double x, double y) override { return _screen_to_touch(x, y); }
+    virtual std::pair<int, int> screen_to_touch(double x, double y) override { return screen_to_touch_no_orient(x, y); }
 
 private:
     template <typename T1, typename T2>
-    inline std::pair<int, int> _screen_to_touch(T1 x, T2 y)
+    inline std::pair<int, int> screen_to_touch_no_orient(T1 x, T2 y)
     {
         return std::make_pair(static_cast<int>(round(x * xscale_)), static_cast<int>(round(y * yscale_)));
     }
 
     bool invoke_and_read_info();
-    void remove_binary();
 
-    std::filesystem::path agent_path_;
     std::string package_name_;
-    std::shared_ptr<InvokeApp> invoke_app_ = std::make_shared<InvokeApp>();
 };
 
 MAA_CTRL_UNIT_NS_END
